% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_statistics.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate Statistics}
\usage{
calculate_statistics(data, stat_type)
}
\arguments{
\item{stat_type}{either the mean, median or standard deviation}
}
\value{
either the mean, median or standard deviation of average
}
\description{
Calculate Statistics
}
\examples{
DRG <- read_csv("DRG_data.csv")

calculate_statistics(DRG, "mean")
calculate_statistics(DRG,"median")
calculate_statistics(DRG,"std")
}
