% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_boxplot.R
\name{drg_boxplot}
\alias{drg_boxplot}
\title{DRG Boxplot (Function 1)}
\usage{
drg_boxplot(
  data,
  variable = c("Average Covered Charges", "Average Total Payments",
    "Average Medicare Payments")
)
}
\arguments{
\item{data}{a dataframe with the DRG data}

\item{variable}{a string name for the variable in the dataframe data}
}
\value{
A boxplot of \code{variable} by DRG code
}
\description{
This function will produce a ggplot boxplot of \code{variable} by DRG code.
}
